+++ src/main/java/com/marcarndt/morse/command/commandlets/chef/config/UpdateUser.java
@@ -1,32 +1,35 @@
-package com.marcarndt.morse.command.commandlets.chefconfig;
+package com.marcarndt.morse.command.commandlets.chef.config;
 
 import com.marcarndt.morse.MorseBot;
-import com.marcarndt.morse.command.ChefConfigure;
 import com.marcarndt.morse.command.commandlet.Commandlet;
+import com.marcarndt.morse.service.ChefService;
 import com.marcarndt.morse.telegrambots.api.objects.Message;
 import java.util.List;
 import javax.ejb.Stateless;
+import javax.inject.Inject;
 
 /**
  * Created by arndt on 2017/05/04.
  */
 @Stateless
-public class Org implements Commandlet {
+public class UpdateUser implements Commandlet {
 
-  public static String ChefOrgStage = "ChefOrg";
+  @Inject
+  ChefService chefService;
 
   public boolean canHandleCommand(Message message, String state) {
-    return state.equals(ChefConfigure.CHEF_CONFIG_STATE) && message.getText()
-        .equals(ChefConfigure.ORG);
+    return state.equals(User.ChefUserState);
   }
 
   public void handleCommand(Message message, String state, List<String> parameters,
       MorseBot morseBot) {
-    morseBot.sendReplyMessage(message, "Enter new ORG name");
+    String user = message.getText();
+    chefService.updateUser(user);
+    morseBot.sendMessage("Updated user", message.getChatId().toString());
   }
 
   public String getNewState(Message message, String command) {
-    return ChefOrgStage;
+    return null;
   }
 
   public List<String> getNewStateParams(Message message, String state, List<String> parameters) {
